chore(ci): hopefully ensure that packages are tested if there are errors
`hermit search -s <pkg>` will ignore manifest errors, deliberately, but as that
was what was used to test new/updated packages, they would be completely skipped.

This PR changes the approach so that the package is always tested at least
once, then search is used to find all available versions for further testing.